keps-383: rescope graduation criterias
The issue for KEP-383 can't be closed although the KEP has graduated to
stable because all the Kubernetes components haven't been migrated to
use the new API. The actual migration is a big effort that needs to be
properly planned and as such would be better handled in a seperate KEP.
There are also some changes that needs to be considered in order to not
break existing behavior. As such, the plan is to remove the migration
from KEP-383 scope and instead do it as part of KEP-3728.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>